some changes in exercise3

+++ Exercicio3/src/Carga.java
@@ -13,7 +13,7 @@ public class Carga {
         return peso;
     }
 
-    public void setPeso(int peso) {
+    public void setPeso(double peso) {
         this.peso = peso;
     }
 
@@ -21,7 +21,7 @@ public class Carga {
         return valor;
     }
 
-    public void setValor(int valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
 }
